Refactor:Mv mod `visitor` `vfs` under `crates_io`
